Fix sandbox proxy PR checks
